requirements.txt
# macOS AI Agent - Dependencies

# Core
requests>=2.31.0

# Screen automation
pyautogui>=0.9.54
pillow>=10.0.0

# Voice
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
pyaudio>=0.2.14

# CLI (optional - for better terminal output)
rich>=13.0.0
